requirements.txt
requests
discord.py